Fix ARM PLT lookups.
An ARM PLT may not contain fixed-length records, because Thumb stubs may
be added to provide interworking.  These changes ensure the breakpoint
gets set on the ARM portion of PLT entries while skipping any Thumb
stubs that may be present.  This patch makes the testsuite work when
built for ARMv7a and Thumb-2.

Signed-off-by: Zachary T Welch <zwelch@codesourcery.com>