Allow building against libelf and elfutils
Adds configure-time check to test the function signature of elf_hash
function, which may be provided by the libelf or elfutils packages.
In the former, the 'name' argument must be an unsigned char pointer.
In the later, the same argument must be a signed char pointer.
The check sets the ELF_HASH_TAKES_SIGNED_CHAR symbol in config.h,
and this is used at the existing call site to pass the proper type.

Signed-off-by: Zachary T Welch <zwelch@codesourcery.com>