Use more warnings by default and make them errors
Compiler warnings should not be limited to building with debugging,
so this patch enbles them for all builds.  Since -Werror may cause
unforseen problems, it provides the --disable-werror option to allow
users to get around any possible problems that it may cause; however,
the proper solution will always be to fix the code to eliminate the
warning from occuring in the first place.

This patch also puts these flags in AM_CFLAGS, so they will not
be affected by the user providing their own CFLAGS settings.

Signed-off-by: Zachary T Welch <zwelch@codesourcery.com>