Remove -O1/-O2/-g from configure script
The default CFLAGS are '-g -O2', so providing these options is either
redundant or conflicting.  The user can specify their own CFLAGS to
override the defaults, so the configure script should not be setting
these itself.  See the documentation for the AC_PROG_CC macro in the
autoconf manual for more information about these defaults.

Signed-off-by: Zachary T Welch <zwelch@codesourcery.com>